Unify on setting type
