+++ sdk/log/provider_test.go
@@ -155,11 +155,6 @@ func TestNewLoggerProviderConfiguration(t *testing.T) {
 	}
 }
 
-func TestLimitValueFailsOpen(t *testing.T) {
-	var l limit
-	assert.Equal(t, -1, l.Value(), "limit value should default to unlimited")
-}
-
 func TestLoggerProviderConcurrentSafe(t *testing.T) {
 	const goRoutineN = 10
